Upload "latest" copies of checksums
